pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikidata-collector"
version = "0.1.0"
description = "Python library for fetching and normalizing public entities from Wikidata via SPARQL"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Jonathan Klüser", email = "jonathan.klueser@live.de"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "random-user-agent>=1.0.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.14.9",
    "pyright>=1.1.390",
]

[tool.pyright]
include = ["wikidata_collector", "tests"]
exclude = ["**/__pycache__", ".venv", "venv", "htmlcov"]
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false

[tool.setuptools.packages.find]
where = ["."]
include = ["wikidata_collector*"]
exclude = ["tests*", "specs*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
